Move walkers to separate module
